Added UI to select rotate or move mode
