+++ 3DAmsterdam/Assets/T3D/Images/ShapeableUI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 177a755eb8ee74454ba54c08c6ff05cd
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
