--- 3DAmsterdam/Assets/T3D/Scripts/UitbouwTransformButton.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2b84030f2bdfd4a388346611900792d4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
